Split out some separate components, move TOC contents to index
