+++ src/components/Article.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Article = ({title}) => {
+  return (
+    <li>{title}</li>
+  )
+}
+
+export default Article;
