+++ src/components/Category.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Category = ({title, children}) => {
+  return (
+    <>
+      <li>{title}</li>
+      <ul>{children}</ul>
+    </>
+  )
+}
+
+export default Category;
